Replace Asyncio_redis with aioredis (#973)
* Initial commit - replace asyncio_redis with aioredis

* Add aioredis to requirements, remove asyncio_redis

* remove asyncio_redis from imports

* Add gcc to docker and then remove after pip install

* Replace old asyncio_redis code to aioredis, update tests

* Move JSONEncoder/Decoder from sqlite to helper

* Remove commented out code

* Handle case where redis is in config but umable to connect
