+++ requirements.txt
@@ -1,9 +1,9 @@
 aiohttp==3.5.4
+aioredis==1.2.0
 aioslacker==0.0.11
 aiosqlite==0.10.0
 appdirs==1.4.3
 arrow==0.14.1
-asyncio_redis==0.15.1
 Babel==2.7.0
 bleach==3.1.0
 click==7.0
