+++ tests/test_database_redis.py
@@ -1,5 +1,5 @@
 import asyncio
-import datetime
+import aioredis
 import unittest
 
 import asynctest
@@ -35,36 +35,7 @@ class TestRedisDatabase(unittest.TestCase):
         self.assertEqual("localhost", database.host)
         self.assertEqual(6379, database.port)
         self.assertEqual(None, database.password)
-
-    def test_other(self):
-        unserialized_data = {
-            "example_string": "test",
-            "example_datetime": datetime.datetime.utcfromtimestamp(1538389815),
-            "example_date": datetime.date.fromtimestamp(1538366400),
-        }
-
-        serialized_data = RedisDatabase.convert_object_to_timestamp(unserialized_data)
-
-        self.assertEqual(serialized_data["example_string"], "test")
-        # Typically I would do assertDictEqual on the result, but as datetime are parsed based on the
-        # timezone of the computer it makes the unittest fragile depending on the timezone of the user.
-        self.assertEqual(serialized_data["example_datetime"][0:10], "datetime::")
-        self.assertEqual(serialized_data["example_date"][0:6], "date::")
-
-    def test_convert_timestamp_to_object(self):
-        serialized_data = {
-            "example_date": "date::1538366400",
-            "example_datetime": "datetime::1538389815",
-            "example_string": "test",
-        }
-
-        unserialized_data = RedisDatabase.convert_timestamp_to_object(serialized_data)
-
-        self.assertEqual(unserialized_data["example_string"], "test")
-        # Typically I would do assertDictEqual on the result, but as datetime are parsed based on the
-        # timezone of the computer it makes the unittest fragile depending on the timezone of the user.
-        self.assertIsInstance(unserialized_data["example_datetime"], datetime.datetime)
-        self.assertIsInstance(unserialized_data["example_date"], datetime.date)
+        self.assertLogs("_LOGGER", "debug")
 
 
 class TestRedisDatabaseAsync(asynctest.TestCase):
@@ -73,21 +44,41 @@ class TestRedisDatabaseAsync(asynctest.TestCase):
     async def test_connect(self):
         opsdroid = amock.CoroutineMock()
         database = RedisDatabase({}, opsdroid=opsdroid)
-        import asyncio_redis
 
-        with amock.patch.object(
-            asyncio_redis.Connection, "create"
-        ) as mocked_connection:
+        with amock.patch.object(aioredis, "create_pool") as mocked_connection:
             mocked_connection.side_effect = NotImplementedError
 
             with suppress(NotImplementedError):
                 await database.connect()
                 self.assertTrue(mocked_connection.called)
+                self.assertLogs("_LOGGER", "info")
+
+    async def test_connect_failure(self):
+        opsdroid = amock.CoroutineMock()
+        database = RedisDatabase({}, opsdroid=opsdroid)
+
+        with amock.patch.object(aioredis, "create_pool") as mocked_connection:
+            mocked_connection.side_effect = OSError()
+
+            with suppress(OSError):
+                await database.connect()
+                self.assertLogs("_LOGGER", "warning")
+
+    async def test_connect_logging(self):
+        opsdroid = amock.CoroutineMock()
+        database = RedisDatabase({}, opsdroid=opsdroid)
+
+        with amock.patch.object(aioredis, "create_pool") as mocked_connection:
+            mocked_connection.set_result = amock.CoroutineMock()
+
+            await database.connect()
+            self.assertTrue(mocked_connection.called)
+            self.assertLogs("_LOGGER", "info")
 
     async def test_get(self):
         db = RedisDatabase({})
         db.client = MockRedisClient()
-        db.client.get = amock.CoroutineMock(return_value='{"key":"value"}')
+        db.client.execute = amock.CoroutineMock(return_value='{"key":"value"}')
 
         result = await db.get("string")
 
@@ -96,7 +87,7 @@ class TestRedisDatabaseAsync(asynctest.TestCase):
     async def test_get_return_None(self):
         db = RedisDatabase({})
         db.client = MockRedisClient()
-        db.client.get = amock.CoroutineMock(return_value=None)
+        db.client.execute = amock.CoroutineMock(return_value=None)
 
         result = await db.get("string")
 
@@ -105,7 +96,7 @@ class TestRedisDatabaseAsync(asynctest.TestCase):
     async def test_put(self):
         db = RedisDatabase({})
         db.client = MockRedisClient()
-        db.client.set = amock.CoroutineMock(return_value='{"key":"value"}')
+        db.client.execute = amock.CoroutineMock(return_value='{"key":"value"}')
 
         result = await db.put("string", dict(key="value"))
